Code Implementation/requirements.txt
google-adk
python-dotenv
uvicorn
google-genai
fastapi
requests